Rename DeployTest as LiveDbTest, move to test_base
Useful outside the deploy context
