+++ tests/test_base.py
@@ -1,8 +1,9 @@
 import orlo
 import unittest
 import os
-from orlo.orm import db
 from flask_testing import TestCase, LiveServerTestCase
+from orlo.util import append_or_create_platforms
+from orlo.orm import db, Release, Package
 
 try:
     TRAVIS = True if os.environ['TRAVIS'] == 'true' else False
@@ -69,6 +70,77 @@ class OrloLiveTest(LiveServerTestCase):
         db.drop_all()
         db.session.remove()
         db.get_engine(self.app).dispose()
+
+
+class LiveDbTest(OrloLiveTest):
+    def setUp(self):
+        # super(DeployTest, self).setUp()
+        db.create_all()
+
+    @staticmethod
+    def _create_release(user='testuser',
+                        team='test team',
+                        platforms=None,
+                        references=None,
+                        success=True):
+        """
+        Create a release using internal methods
+
+        :param user:
+        :param team:
+        :param platforms:
+        :param references:
+        :return:
+        """
+
+        if not platforms:
+            platforms = ['test_platform']
+        if type(platforms) is not list:
+            raise AssertionError("Platforms parameter must be list")
+        if not references:
+            references = ['TestTicket-123']
+
+        db_platforms = append_or_create_platforms(platforms)
+
+        r = Release(
+            platforms=db_platforms,
+            user=user,
+            references=references,
+            team=team,
+        )
+        db.session.add(r)
+        db.session.commit()
+
+        return r.id
+
+    @staticmethod
+    def _create_package(release_id,
+                        name='test-package',
+                        version='1.2.3',
+                        diff_url=None,
+                        rollback=False,
+                        ):
+        """
+        Create a package using internal methods
+
+        :param release_id:
+        :param name:
+        :param version:
+        :param diff_url:
+        :param rollback:
+        :return:
+        """
+        p = Package(
+            release_id=release_id,
+            name=name,
+            version=version,
+            diff_url=diff_url,
+            rollback=rollback,
+        )
+        db.session.add(p)
+        db.session.commit()
+
+        return p.id
 
 
 class ConfigChange(object):
